Update README.md
Trying to add some extra explanation to the scheme. Unfortunately, the table does not (yet?) show that D is overwritten every 16th backup (it looks to have the same cadence as C).
Besides that, it seems that when A was the most recent backup, you can restore your file from 1, 2, 4, or 8 backups ago. But when B was the most recent backup, you can restore your file from 1, 2, 3, 7 backups ago. So this part may need a bit extra explanation as well.
Kind regards,
Roger